requirements.txt
altgraph==0.17.4
click==8.3.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
macholib==1.16.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
tk==0.1.0
